feat(rust): complete delete function by enforcing TTY --yes flag

+++ implementations/rust/ockam/ockam_command/src/space/delete.rs
@@ -1,10 +1,12 @@
 use clap::Args;
 use colorful::Colorful;
+use miette::miette;
 
 use ockam::Context;
 use ockam_api::cli_state::{StateDirTrait, StateItemTrait};
 
 use crate::node::util::{delete_embedded_node, start_embedded_node};
+use crate::terminal::ConfirmResult;
 use crate::util::api::{self, CloudOpts};
 use crate::util::{node_rpc, RpcBuilder};
 use crate::{docs, fmt_ok, CommandGlobalOpts};
@@ -26,6 +28,10 @@ pub struct DeleteCommand {
 
     #[command(flatten)]
     pub cloud_opts: CloudOpts,
+
+    /// Confirm the deletion without prompting
+    #[arg(display_order = 901, long, short)]
+    yes: bool,
 }
 
 impl DeleteCommand {
@@ -53,16 +59,43 @@ async fn run_impl(
 
     // Send request
     let mut rpc = RpcBuilder::new(ctx, &opts, &node_name).build();
-    rpc.request(api::space::delete(&id, controller_route))
-        .await?;
-    rpc.is_ok()?;
+    if cmd.yes {
+        rpc.request(api::space::delete(&id, controller_route))
+            .await?;
+        rpc.is_ok()?;
 
-    // Remove from state
-    let _ = opts.state.spaces.delete(&cmd.name);
-    // TODO: remove projects associated to the space.
-    //  Currently we are not storing that association in the project config file.
+        // Remove from state
+        let _ = opts.state.spaces.delete(&cmd.name);
+        // TODO: remove projects associated to the space.
+        //  Currently we are not storing that association in the project config file.
 
-    delete_embedded_node(&opts, rpc.node_name()).await;
+        delete_embedded_node(&opts, rpc.node_name()).await;
+    } else {
+        // If yes is not provided make sure using TTY
+        match opts
+            .terminal
+            .confirm("This will delete the selected Space. Are you sure?")?
+        {
+            ConfirmResult::Yes => {
+                rpc.request(api::space::delete(&id, controller_route))
+                    .await?;
+                rpc.is_ok()?;
+
+                // Remove from state
+                let _ = opts.state.spaces.delete(&cmd.name);
+                // TODO: remove projects associated to the space.
+                //  Currently we are not storing that association in the project config file.
+
+                delete_embedded_node(&opts, rpc.node_name()).await;
+            }
+            ConfirmResult::No => {
+                return Ok(());
+            }
+            ConfirmResult::NonTTY => {
+                return Err(miette!("Use --yes to confirm").into());
+            }
+        }
+    }
 
     // log the deletion
     opts.terminal
